fix(api) fix concurrency issues of PrismaEventRepository
Replace Interactive Transactions with standard $transaction API.
Ensure consistiency of aggragates using additional table EventLock with
unique constraints on (streamId, streamVersion)
