+++ packages/api/prisma/schema.prisma
@@ -4,8 +4,7 @@ datasource db {
 }
 
 generator client {
-  provider        = "prisma-client-js"
-  previewFeatures = ["interactiveTransactions"]
+  provider = "prisma-client-js"
 }
 
 model UserProfile {
@@ -68,4 +67,11 @@ model EventsSubscription {
   currentPosition Int
   fromPosition    Int
   eventTypes      String[]
+}
+
+model EventLock {
+  streamId      String
+  streamVersion Int
+
+  @@unique([streamId, streamVersion])
 }
